+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 0.6.9 Release (03/23/2026)
+- Added new `-an` option to export markdown in a fully compatible format with Apple Notes markdown file import
+- Added new `-h` option to dynamically change hashtags to labels in existing Keep notes (experimental - up to 5 at a time)
+- Added special instructions for Apple Notes and Hashtag conversions
 
 ## 0.6.8 Release (08/13/2025)
 - Fixed bug using `-m` option to move notes to archive with the `--all` flag when filtering notes by create or edit dates (all notes were being moved vs only those filtered)